[Reafac] RoundReult Scene, GameResult Scene전환 로직을 Scene Initializer 스크립트로 변경
