--- JKC_Fallguys/Assets/1_Scripts/UI/05_RoundResult/SceneChanger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8b921f8fdc18b48e5ad4d4e1a35f9422
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
