--- JKC_Fallguys/Assets/1_Scripts/etc/06_GameResult/AnimationEventController.cs
@@ -1,11 +1,0 @@
-using LiteralRepository;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class AnimationEventController : MonoBehaviour
-{
-    public void LoadLobbyScene()
-    {
-        SceneManager.LoadScene(SceneIndex.Lobby);
-    }
-}
