--- JKC_Fallguys/Assets/1_Scripts/etc/06_GameResult/AnimationEventController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 632ca4e385a224cd8a67407b26ac9904
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
