--- JKC_Fallguys/Assets/Resources/Character/Animation/FG_Idle_Qualification01.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4203d3ab5c6454e2eb5b91be69580c29
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
