--- JKC_Fallguys/Assets/Resources/Prefabs/UI/05_RoundResult/SceneChanger.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 057f1637aa329483394e649e05f52064
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
