+++ JKC_Fallguys/ProjectSettings/EditorBuildSettings.asset
@@ -26,10 +26,4 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/2_Scenes/06_GameResult.unity
     guid: 5d9eb69ddf6f345f39791ac99516fac9
-  - enabled: 0
-    path: Assets/Jinsoo/TestStage.unity
-    guid: c780a2c1e7a284bdba13dbfde013ac7f
-  - enabled: 0
-    path: Assets/Huiji/Scenes/StageUI.unity
-    guid: 1027bc7b0357c496f9130abc318c2b09
   m_configObjects: {}
